pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-egress-guard"
version = "0.1.0"
description = "A deterministic egress guard for LLM responses"
readme = "README.md"
authors = [{name = "LLM Egress Guard Team"}]
license = {file = "LICENSE"}
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115,<1",
  "uvicorn>=0.30,<1",
  "pydantic>=2.9,<3",
  "pydantic-settings>=2.5,<3",
  "pyyaml>=6.0,<7",
  "structlog>=24.1,<25",
  "python-json-logger>=2.0,<3",
  "orjson>=3.10,<4",
  "prometheus-client>=0.20,<1"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3,<9",
  "pytest-asyncio>=0.23,<1",
  "httpx>=0.27,<0.29",
  "black>=24.10,<25",
  "ruff>=0.6,<0.7"
]

[tool.pytest.ini_options]
addopts = "-ra"
minversion = "7.0"
pythonpath = ["app"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
src = ["app", "tests"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "N"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
